chore: modernize the inflect package
The inflect package is a discreet yet real dependency for go-swagger
(the pluralize feature is consumed by the client codegen template).

In the future, plurarizing words could be used in other name mangling utilities
(processing "initialisms" in go-openapi/swag).

This PR makes sure that the go-openapi/inflect package stays aligned
with the other go-openapi packages in terms of linting, documentation
and testing.

Signed-off-by: Frédéric BIDON <fredbi@yahoo.com>